[SCRUM-150] Bug : 오타 수정..
