+++ src/main/java/com/kkinikong/be/community/controller/CommunityController.java
@@ -1,4 +1,4 @@
-package com.kkinikong.be.commuity.controller;
+package com.kkinikong.be.community.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
