+++ src/main/java/com/kkinikong/be/community/domain/Comment.java
@@ -1,5 +1,9 @@
-package com.kkinikong.be.commuity.domain;
+package com.kkinikong.be.community.domain;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,6 +12,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,16 +21,28 @@ import lombok.NoArgsConstructor;
 import com.kkinikong.be.global.entity.BaseEntity;
 import com.kkinikong.be.user.domain.User;
 
-@Table(name = "community_post_likes")
+@Table(name = "comments")
 @Entity
 @Getter
 @NoArgsConstructor
-public class CommunityPostLike extends BaseEntity {
+public class Comment extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "id")
   private Long id;
+
+  @Column(name = "content", nullable = false, length = 4000)
+  private String content;
+
+  @Column(name = "like_count", nullable = false)
+  private long likeCount = 0L;
+
+  @Column(name = "is_deleted", nullable = false)
+  private boolean isDeleted = false;
+
+  @Column(name = "parent_comment_id")
+  private Long parentCommentId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id", nullable = false)
@@ -35,9 +52,14 @@ public class CommunityPostLike extends BaseEntity {
   @JoinColumn(name = "community_post_id", nullable = false)
   private CommunityPost communityPost;
 
+  @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL, orphanRemoval = true)
+  private List<CommentLike> commentLikeList = new ArrayList<>();
+
   @Builder
-  public CommunityPostLike(User user, CommunityPost communityPost) {
+  public Comment(String content, User user, CommunityPost communityPost, Long parentCommentId) {
+    this.content = content;
     this.user = user;
     this.communityPost = communityPost;
+    this.parentCommentId = parentCommentId;
   }
 }
