+++ src/main/java/com/kkinikong/be/community/domain/CommunityPostImage.java
@@ -1,4 +1,4 @@
-package com.kkinikong.be.commuity.domain;
+package com.kkinikong.be.community.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -14,30 +14,28 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import com.kkinikong.be.global.entity.BaseEntity;
-import com.kkinikong.be.user.domain.User;
 
-@Table(name = "comment_likes")
+@Table(name = "community_post_images")
 @Entity
 @Getter
 @NoArgsConstructor
-public class CommentLike extends BaseEntity {
+public class CommunityPostImage extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "id")
   private Long id;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
+  @Column(name = "image_url", nullable = false)
+  private String imageUrl;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "comment_id", nullable = false)
-  private Comment comment;
+  @JoinColumn(name = "community_post_id", nullable = false)
+  private CommunityPost communityPost;
 
   @Builder
-  public CommentLike(User user, Comment comment) {
-    this.user = user;
-    this.comment = comment;
+  public CommunityPostImage(String imageUrl, CommunityPost communityPost) {
+    this.imageUrl = imageUrl;
+    this.communityPost = communityPost;
   }
 }
