+++ src/main/java/com/kkinikong/be/community/domain/CommunityPostLike.java
@@ -1,9 +1,5 @@
-package com.kkinikong.be.commuity.domain;
+package com.kkinikong.be.community.domain;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -12,7 +8,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Builder;
 import lombok.Getter;
@@ -21,28 +16,16 @@ import lombok.NoArgsConstructor;
 import com.kkinikong.be.global.entity.BaseEntity;
 import com.kkinikong.be.user.domain.User;
 
-@Table(name = "comments")
+@Table(name = "community_post_likes")
 @Entity
 @Getter
 @NoArgsConstructor
-public class Comment extends BaseEntity {
+public class CommunityPostLike extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "id")
   private Long id;
-
-  @Column(name = "content", nullable = false, length = 4000)
-  private String content;
-
-  @Column(name = "like_count", nullable = false)
-  private long likeCount = 0L;
-
-  @Column(name = "is_deleted", nullable = false)
-  private boolean isDeleted = false;
-
-  @Column(name = "parent_comment_id")
-  private Long parentCommentId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id", nullable = false)
@@ -52,14 +35,9 @@ public class Comment extends BaseEntity {
   @JoinColumn(name = "community_post_id", nullable = false)
   private CommunityPost communityPost;
 
-  @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL, orphanRemoval = true)
-  private List<CommentLike> commentLikeList = new ArrayList<>();
-
   @Builder
-  public Comment(String content, User user, CommunityPost communityPost, Long parentCommentId) {
-    this.content = content;
+  public CommunityPostLike(User user, CommunityPost communityPost) {
     this.user = user;
     this.communityPost = communityPost;
-    this.parentCommentId = parentCommentId;
   }
 }
