+++ src/main/java/com/kkinikong/be/community/repository/CommentRepository.java
@@ -1,9 +1,9 @@
-package com.kkinikong.be.commuity.repository;
+package com.kkinikong.be.community.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.kkinikong.be.commuity.domain.Comment;
+import com.kkinikong.be.community.domain.Comment;
 
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {}
