+++ src/main/java/com/kkinikong/be/community/repository/CommunityPostImageRepository.java
@@ -1,9 +1,9 @@
-package com.kkinikong.be.commuity.repository;
+package com.kkinikong.be.community.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.kkinikong.be.commuity.domain.CommunityPostImage;
+import com.kkinikong.be.community.domain.CommunityPostImage;
 
 @Repository
 public interface CommunityPostImageRepository extends JpaRepository<CommunityPostImage, Long> {}
