+++ src/main/java/com/kkinikong/be/community/repository/CommunityPostLikeRepository.java
@@ -1,9 +1,9 @@
-package com.kkinikong.be.commuity.repository;
+package com.kkinikong.be.community.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.kkinikong.be.commuity.domain.CommunityPostLike;
+import com.kkinikong.be.community.domain.CommunityPostLike;
 
 @Repository
 public interface CommunityPostLikeRepository extends JpaRepository<CommunityPostLike, Long> {}
