+++ src/main/java/com/kkinikong/be/community/repository/CommunityRepository.java
@@ -1,9 +1,9 @@
-package com.kkinikong.be.commuity.repository;
+package com.kkinikong.be.community.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.kkinikong.be.commuity.domain.CommunityPost;
+import com.kkinikong.be.community.domain.CommunityPost;
 
 @Repository
 public interface CommunityRepository extends JpaRepository<CommunityPost, Long> {}
